src/requirements.txt
pandas==2.1.3
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.29.0
scikit-learn==1.3.2